Removed VPC firewall policy association in GCP
In all of the other clouds we don't do the association. (And in
AWS it's not even possible.)

This commit isn't being squashed though so that the association
code is in the git history if we decide we want it later and
don't want to have to redo the work relating to learning how
to make the association. (Though we do need to get the full
VPC URL, which isn't in the code.)